Fixed Views Counter
No views will be counted for admin.

+++ inc/template-functions.php
@@ -216,15 +216,17 @@ function getPostViews($postID){
     return $count.' Views';
 }
 function setPostViews($postID) {
-    $count_key = 'post_views_count';
-    $count = get_post_meta($postID, $count_key, true);
-    if($count==''){
-        $count = 0;
-        delete_post_meta($postID, $count_key);
-        add_post_meta($postID, $count_key, '0');
-    }else{
-        $count++;
-        update_post_meta($postID, $count_key, $count);
+    if ( !is_admin_bar_showing() && !is_customize_preview() ) {
+        $count_key = 'post_views_count';
+        $count = get_post_meta($postID, $count_key, true);
+        if ( $count=='' ) {
+            $count = 0;
+            delete_post_meta($postID, $count_key);
+            add_post_meta($postID, $count_key, '0');
+        } else {
+            $count++;
+            update_post_meta($postID, $count_key, $count);
+        }
     }
 }
 // Remove issues with prefetching adding extra views
